pyproject.toml
[project]
name = "tomojlifeproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.5.2",
    "fastapi>=0.135.3",
    "langchain-openai>=1.1.12",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
